services/vanna/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
vanna
python-dotenv
groq
